some scaling improvements

+++ in4073/pc_terminal/packet_pc.c
@@ -18,10 +18,10 @@ void encode(packet *my_packet, uint8_t mode, int16_t roll_r, int16_t pitch_r, in
          */
          int8_t roll_c, pitch_c, yaw_c;
          uint8_t lift_c;
-         roll_c = ((float)roll_r/32768)*127;
-         pitch_c =((float)pitch_r/32768)*127;
-         yaw_c = ((float)yaw_r/32768)*127;
-         lift_c = (((float)lift_r+32768)/65536)*255;
+         roll_c = (int8_t)(((float)roll_r/32768)*127);
+         pitch_c =(int8_t)(((float)pitch_r/32768)*127);
+         yaw_c = (int8_t)(((float)yaw_r/32768)*127);
+         lift_c = (uint8_t)((((float)lift_r+32768)/65536)*255);
          //printf("roll_c: %x  roll_r : %x\n",roll_c, roll_r );
         /*
            Construct the Packet
